Added scanner utility script
This will help identify what references to Neutron are made by code in
a repo (e.g. LBaaS, FWaaS, VPNaaS). It analyzes imports and then looks
at the use of the import in source modules.

Run this script using python, and supply the root of the directory
tree to analyze.