Revised scanner to handle imports of directories
Found that there were some cases where directories were imported, instead
of modules (or objects). Adjusted scanner to handle this, added a test
case, and run on recent copies of FW, LB, and VPN repos.

Included both the full analysis and just the summary for each of the three
repos.

Move everything for scanner into a sub-directory to separate from other
tools.